requirements.txt
asgiref==3.5.1
Django==4.0.4
django-rest-auth==0.9.5
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
